Split StanCompiler and StanRunner.
