+++ src/main/scala/com/cibo/scalastan/run/StanRunner.scala
@@ -8,21 +8,16 @@
  * or at https://opensource.org/licenses/BSD-3-Clause
  */
 
-package com.cibo.scalastan
+package com.cibo.scalastan.run
 
-trait StanRunner[M <: CompiledModel] {
-  def compile(ss: ScalaStan, model: ScalaStan#Model): CompiledModel
+import com.cibo.scalastan.{CompiledModel, RunMethod, StanResults}
+
+trait StanRunner {
   def run(
-    model: M,
+    model: CompiledModel,
     chains: Int,
     seed: Int,
     cache: Boolean,
     method: RunMethod.Method
   ): StanResults
-}
-
-object StanRunner {
-
-  // Use CmdStan to compile and run the model.
-  implicit val defaultRunner: StanRunner[CmdStanCompiledModel] = CmdStanRunner
 }
